sitemap generating for ready category seo moved to framework package
