+++ packages/framework/src/Model/Sitemap/SitemapFacade.php
@@ -99,4 +99,13 @@ class SitemapFacade
     {
         return $this->sitemapRepository->getSitemapItemsForVisibleFlags($domainConfig);
     }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
+     * @return \Shopsys\FrameworkBundle\Model\Sitemap\SitemapItem[]
+     */
+    public function getSitemapItemsForVisibleCategorySeoMix(DomainConfig $domainConfig): array
+    {
+        return $this->sitemapRepository->getSitemapItemsForVisibleCategorySeoMix($domainConfig);
+    }
 }
